消费者模块搭建完成，主要： （1）consumer项目模拟用户操作，调取paymeny服务接口 （2）调用方式选择的是，使用spring 的 RestTemplate工具 （3）Controller 记得加上@RequestBody注解

+++ cloud-provider-payment8001/src/main/java/com/atguigu/springcloud/dao/PaymentMapper.java
@@ -9,5 +9,5 @@ public interface PaymentMapper {
 
     int save(Payment payment);
 
-    Payment getPaymentById(@Param("id") Integer id);
+    Payment getPaymentById(@Param("id") Long id);
 }
