+++ cloud-provider-payment8001/src/main/java/com/atguigu/springcloud/service/PaymentService.java
@@ -6,5 +6,5 @@ import org.apache.ibatis.annotations.Param;
 public interface PaymentService {
     int save(Payment payment);
 
-    Payment getPaymentById(@Param("id") Integer id) ;
+    Payment getPaymentById(@Param("id") Long id) ;
 }
